pyproject.toml
[project]
name = "quicketl"
version = "1.2.0"
description = "QuickETL - Fast & Flexible Python ETL Framework with 20+ backend support via Ibis"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [
    { name = "Eiji", email = "eiji@eidosoft.co" }
]
keywords = ["etl", "elt", "data-engineering", "ibis", "polars", "duckdb", "pipeline", "quicketl"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database",
    "Topic :: Scientific/Engineering",
    "Typing :: Typed",
]

dependencies = [
    # Core
    "pydantic>=2.10",
    "pyyaml>=6.0",
    "structlog>=24.0",
    "typer>=0.12",
    "rich>=13.0",

    # Ibis (the engine abstraction layer) - DuckDB and Polars by default
    "ibis-framework[duckdb,polars]>=9.0",

    # Cloud storage abstraction
    "fsspec>=2024.6",
]

[project.optional-dependencies]
# Additional Ibis backends
spark = [
    "ibis-framework[pyspark]>=9.0",
    "setuptools>=60.0",  # Provides distutils shim for Python 3.12+
]
pandas = ["pandas>=2.0"]  # pandas support is built into ibis-framework
datafusion = ["ibis-framework[datafusion]>=9.0"]

# Cloud storage providers
aws = ["s3fs>=2024.6", "boto3>=1.34"]
gcp = ["gcsfs>=2024.6", "google-cloud-storage>=2.14"]
azure = ["adlfs>=2024.4", "azure-storage-blob>=12.19"]

# Data warehouses (via Ibis)
snowflake = ["ibis-framework[snowflake]>=9.0"]
bigquery = ["ibis-framework[bigquery]>=9.0"]
trino = ["ibis-framework[trino]>=9.0"]
postgres = ["ibis-framework[postgres]>=9.0"]
mysql = ["ibis-framework[mysql]>=9.0"]
clickhouse = ["ibis-framework[clickhouse]>=9.0"]

# Secrets providers
secrets-aws = ["boto3>=1.34"]
secrets-azure = ["azure-identity>=1.15", "azure-keyvault-secrets>=4.8"]

# AI/ML - Text chunking
chunking = ["tiktoken>=0.5", "nltk>=3.8"]

# AI/ML - Embeddings
embeddings-openai = ["openai>=1.0", "tiktoken>=0.5"]
embeddings-huggingface = ["sentence-transformers>=2.2"]

# Vector stores
vector-pinecone = ["pinecone-client>=3.0"]
vector-pgvector = ["pgvector>=0.2", "psycopg2-binary>=2.9"]
vector-qdrant = ["qdrant-client>=1.7"]

# Telemetry
opentelemetry = [
    "opentelemetry-api>=1.22",
    "opentelemetry-sdk>=1.22",
    "opentelemetry-exporter-otlp>=1.22",
]
openlineage = ["openlineage-python>=1.8"]

# Development tools
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "hypothesis>=6.100",
    "ruff>=0.5",
    "mypy>=1.10",
    "pre-commit>=3.7",
]

# Documentation
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.25",
    "mkdocs-gen-files>=0.5",
    "mkdocs-literate-nav>=0.6",
    "mkdocs-section-index>=0.3",
    "pymdown-extensions>=10.0",
]

# AI bundle (chunking, embeddings, vector stores)
ai = [
    "quicketl[chunking,embeddings-openai,embeddings-huggingface,vector-pinecone,vector-pgvector,vector-qdrant]",
]

# Enterprise bundle (secrets, telemetry)
enterprise = [
    "quicketl[secrets-aws,secrets-azure,opentelemetry,openlineage]",
]

# All backends and tools
all = [
    "quicketl[spark,pandas,datafusion,aws,gcp,azure,snowflake,bigquery,trino,postgres,mysql,clickhouse,ai,enterprise,dev,docs]",
]

[project.scripts]
quicketl = "quicketl.cli.main:cli"

[project.urls]
Homepage = "https://quicketl.com"
Documentation = "https://quicketl.com"
Repository = "https://github.com/ameijin/quicketl"
Issues = "https://github.com/ameijin/quicketl/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
license-files = []

[tool.hatch.build.targets.wheel]
packages = ["src/quicketl"]

[tool.hatch.version]
path = "src/quicketl/_version.py"

# Ruff configuration
[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults (needed for typer)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "ARG001",  # Unused function argument (fixtures often unused)
    "ARG002",  # Unused method argument (fixtures often unused)
    "ARG005",  # Unused lambda argument (mock patterns)
    "TCH002",  # Move third-party import into type-checking block (used at runtime in tests)
    "TCH003",  # Move standard library import into type-checking block (used at runtime in tests)
]

[tool.ruff.lint.isort]
known-first-party = ["quicketl"]

# MyPy configuration
[tool.mypy]
python_version = "3.12"
strict = false
warn_return_any = false
warn_unused_ignores = true
plugins = ["pydantic.mypy"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "ibis.*",
    "fsspec.*",
    "s3fs.*",
    "gcsfs.*",
    "adlfs.*",
    "yaml.*",
    "pandas.*",
]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--cov=quicketl",
    "--cov-report=term-missing",
]
markers = [
    "slow: marks tests as slow (> 5 seconds)",
    "integration: marks integration tests (requires external services)",
    "parity: marks backend parity tests (verifies backend consistency)",
    "hypothesis: marks property-based tests (fuzzing for edge cases)",
]

# Coverage configuration
[tool.coverage.run]
source = ["src/quicketl"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@abstractmethod",
]

# Semantic Release configuration
[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
commit_parser = "conventional"

[tool.semantic_release.changelog]
changelog_file = "CHANGELOG.md"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "twine>=6.2.0",
    "python-semantic-release>=10.0.0",
]
